+++ packages/shadstack-table/src/components/head/SST_TableHeadCellFilterContainer.tsx
@@ -1,4 +1,5 @@
 import * as React from 'react';
+import { useSST_TableState } from '../../hooks/useSST_TableState';
 import { Collapsible, CollapsibleContent } from '../../_ui/collapsible';
 import { type SST_Header, type SST_RowData, type SST_TableInstance } from '../../types';
 import { getColumnFilterInfo } from '../../utils/column.utils';
@@ -20,10 +21,14 @@ export const SST_TableHeadCellFilterContainer = <TData extends SST_RowData>({
   ...rest
 }: SST_TableHeadCellFilterContainerProps<TData>) => {
   const {
-    getState,
     options: { columnFilterDisplayMode },
   } = table;
-  const { showColumnFilters } = getState();
+  const { showColumnFilters } = useSST_TableState(table, (s) => ({
+    showColumnFilters: s.showColumnFilters,
+    //the rendered filter input reflects this column's value and fn
+    filterValue: header.column.getFilterValue(),
+    filterFn: s.columnFilterFns[header.column.id],
+  }));
   const { column } = header;
   const { columnDef } = column;
   const { isRangeFilter } = getColumnFilterInfo({ header, table });